[#180] rename: 패키지명 notification으로 변경
