+++ src/main/java/com/konggogi/veganlife/notification/domain/Notification.java
@@ -1,4 +1,4 @@
-package com.konggogi.veganlife.sse.domain;
+package com.konggogi.veganlife.notification.domain;
 
 
 import com.konggogi.veganlife.global.domain.TimeStamped;
