+++ src/main/java/com/konggogi/veganlife/notification/domain/NotificationMessage.java
@@ -1,4 +1,4 @@
-package com.konggogi.veganlife.sse.domain;
+package com.konggogi.veganlife.notification.domain;
 
 public enum NotificationMessage implements MessageFormatter {
     // sse
