+++ src/main/java/com/konggogi/veganlife/notification/domain/NotificationType.java
@@ -1,4 +1,4 @@
-package com.konggogi.veganlife.sse.domain;
+package com.konggogi.veganlife.notification.domain;
 
 public enum NotificationType {
     SSE,
