+++ src/main/java/com/konggogi/veganlife/notification/repository/EmitterRepository.java
@@ -1,4 +1,4 @@
-package com.konggogi.veganlife.sse.repository;
+package com.konggogi.veganlife.notification.repository;
 
 
 import java.util.Optional;
