+++ src/main/java/com/konggogi/veganlife/notification/repository/EmitterRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.konggogi.veganlife.sse.repository;
+package com.konggogi.veganlife.notification.repository;
 
 
 import java.util.Map;
